# web/openapi.yaml
components:
  schemas:
    BootstrapOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/BootstrapOutputBody.json
          format: uri
          readOnly: true
          type: string
        setup_required:
          type: boolean
      required:
        - setup_required
      type: object
    BreakdownItem:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        key:
          type: string
        percent:
          format: double
          type: number
      required:
        - key
        - count
        - percent
      type: object
    BreakdownOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/BreakdownOutputBody.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/BreakdownItem"
          nullable: true
          type: array
        total:
          format: int64
          type: integer
      required:
        - items
        - total
      type: object
    CollectInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CollectInputBody.json
          format: uri
          readOnly: true
          type: string
        event:
          maxLength: 100
          type: string
        hostname:
          maxLength: 255
          type: string
        name:
          default: pageview
          enum:
            - pageview
            - event
          type: string
        path:
          maxLength: 2048
          type: string
        referrer:
          maxLength: 2048
          type: string
        session_id:
          maxLength: 100
          minLength: 8
          type: string
        timestamp:
          format: date-time
          type: string
        title:
          maxLength: 500
          type: string
        visitor_id:
          maxLength: 100
          minLength: 8
          type: string
        website_id:
          format: uuid
          type: string
      required:
        - website_id
        - visitor_id
        - session_id
        - name
        - path
        - hostname
        - title
        - referrer
      type: object
    CollectOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CollectOutputBody.json
          format: uri
          readOnly: true
          type: string
        accepted:
          type: boolean
      required:
        - accepted
      type: object
    CreateWebsiteInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CreateWebsiteInputBody.json
          format: uri
          readOnly: true
          type: string
        domain:
          maxLength: 255
          minLength: 1
          type: string
        name:
          maxLength: 100
          minLength: 1
          type: string
        timezone:
          default: UTC
          maxLength: 100
          type: string
      required:
        - name
        - domain
        - timezone
      type: object
    CredentialsInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CredentialsInputBody.json
          format: uri
          readOnly: true
          type: string
        email:
          format: email
          maxLength: 320
          type: string
        password:
          maxLength: 1024
          minLength: 12
          type: string
      required:
        - email
        - password
      type: object
    Dashboard:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Dashboard.json
          format: uri
          readOnly: true
          type: string
        average_duration:
          $ref: "#/components/schemas/KPI"
        bounce_rate:
          $ref: "#/components/schemas/KPI"
        has_received_event:
          type: boolean
        online_visitors:
          format: int64
          type: integer
        page_view_series:
          items:
            $ref: "#/components/schemas/Point"
          nullable: true
          type: array
        page_views:
          $ref: "#/components/schemas/KPI"
        unique_visitors:
          $ref: "#/components/schemas/KPI"
        views_per_visit:
          $ref: "#/components/schemas/KPI"
        visit_series:
          items:
            $ref: "#/components/schemas/Point"
          nullable: true
          type: array
        visitor_series:
          items:
            $ref: "#/components/schemas/Point"
          nullable: true
          type: array
        visits:
          $ref: "#/components/schemas/KPI"
      required:
        - visits
        - views_per_visit
        - bounce_rate
        - average_duration
        - unique_visitors
        - page_views
        - visit_series
        - visitor_series
        - page_view_series
        - online_visitors
        - has_received_event
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          nullable: true
          type: array
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          example: https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          example: 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          example: Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          example: https://example.com/errors/example
          format: uri
          type: string
      type: object
    KPI:
      additionalProperties: false
      properties:
        change:
          format: double
          type: number
        value:
          format: double
          type: number
      required:
        - value
        - change
      type: object
    LogoutOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/LogoutOutputBody.json
          format: uri
          readOnly: true
          type: string
        ok:
          type: boolean
      required:
        - ok
      type: object
    Point:
      additionalProperties: false
      properties:
        timestamp:
          format: date-time
          type: string
        value:
          format: int64
          type: integer
      required:
        - timestamp
        - value
      type: object
    User:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/User.json
          format: uri
          readOnly: true
          type: string
        email:
          type: string
        id:
          type: string
      required:
        - id
        - email
      type: object
    Visitor:
      additionalProperties: false
      properties:
        browser:
          type: string
        city:
          type: string
        country:
          type: string
        device:
          type: string
        first_seen_at:
          format: date-time
          type: string
        id:
          type: string
        last_active_at:
          format: date-time
          type: string
        name:
          type: string
        os:
          type: string
        referrer:
          type: string
        views:
          format: int64
          type: integer
      required:
        - id
        - name
        - country
        - city
        - referrer
        - device
        - browser
        - os
        - first_seen_at
        - last_active_at
        - views
      type: object
    VisitorsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/VisitorsOutputBody.json
          format: uri
          readOnly: true
          type: string
        total:
          format: int64
          type: integer
        visitors:
          items:
            $ref: "#/components/schemas/Visitor"
          nullable: true
          type: array
      required:
        - visitors
        - total
      type: object
    Website:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Website.json
          format: uri
          readOnly: true
          type: string
        created_at:
          format: date-time
          type: string
        domain:
          type: string
        id:
          type: string
        name:
          type: string
        timezone:
          type: string
      required:
        - id
        - name
        - domain
        - timezone
        - created_at
      type: object
    WebsitesOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/WebsitesOutputBody.json
          format: uri
          readOnly: true
          type: string
        websites:
          items:
            $ref: "#/components/schemas/Website"
          nullable: true
          type: array
      required:
        - websites
      type: object
info:
  description: The local API used by the Dottie dashboard and browser tracker.
  title: Dottie API
  version: dev
openapi: 3.0.3
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Auth
  /api/v1/auth/logout:
    post:
      operationId: logout
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutOutputBody"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Auth
  /api/v1/auth/me:
    get:
      operationId: get-current-user
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Auth
  /api/v1/auth/setup:
    post:
      operationId: setup-admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Auth
  /api/v1/collect:
    post:
      operationId: collect-event
      parameters:
        - in: header
          name: Origin
          schema:
            type: string
        - in: header
          name: User-Agent
          schema:
            type: string
        - in: header
          name: DNT
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectOutputBody"
          description: OK
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Vary:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Collector
  /api/v1/system/bootstrap:
    get:
      operationId: get-bootstrap-status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - System
  /api/v1/websites:
    get:
      operationId: list-websites
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsitesOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Websites
    post:
      operationId: create-website
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebsiteInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Websites
  /api/v1/websites/{website_id}/breakdowns:
    get:
      operationId: get-breakdowns
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
        - in: path
          name: website_id
          required: true
          schema:
            format: uuid
            type: string
        - explode: false
          in: query
          name: period
          schema:
            default: 30d
            enum:
              - 7d
              - 30d
              - 90d
            type: string
        - explode: false
          in: query
          name: dimension
          schema:
            default: pages
            enum:
              - pages
              - referrers
              - countries
              - cities
              - devices
              - browsers
              - os
              - events
            type: string
        - explode: false
          in: query
          name: search
          schema:
            maxLength: 200
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 9
            format: int64
            maximum: 100
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Analytics
  /api/v1/websites/{website_id}/dashboard:
    get:
      operationId: get-dashboard
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
        - in: path
          name: website_id
          required: true
          schema:
            format: uuid
            type: string
        - explode: false
          in: query
          name: period
          schema:
            default: 30d
            enum:
              - 7d
              - 30d
              - 90d
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Analytics
  /api/v1/websites/{website_id}/visitors:
    get:
      operationId: list-visitors
      parameters:
        - in: cookie
          name: dottie_session
          schema:
            type: string
        - in: path
          name: website_id
          required: true
          schema:
            format: uuid
            type: string
        - explode: false
          in: query
          name: search
          schema:
            maxLength: 200
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 15
            format: int64
            maximum: 100
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorsOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Analytics